@@ -1,13 +1,21 @@
 import "./App.scss";
 import "./Responsive.css";
 
+import { makeStyles } from "@mui/styles"
+import { ThemeProvider, createTheme } from '@mui/material/styles';
 import NotificationProvider from "./contexts/NotificationContext";
 import { ProfileProvider } from "./contexts/ProfileContext";
 import React, { useEffect } from "react";
 
 import Routes from "./routes";
 
+const theme = createTheme();
+
+const useStyles = makeStyles();
+
 export default function App() {
+	useStyles();
+
 	useEffect(() => {
 		const usesDarkMode =
 			window.matchMedia("(prefers-color-scheme: dark)").matches || false;
@@ -24,10 +32,12 @@ export default function App() {
 	}, []);
 
 	return (
-		<NotificationProvider>
-			<ProfileProvider>
-				<Routes />
-			</ProfileProvider>
-		</NotificationProvider>
+		<ThemeProvider theme={theme}>
+			<NotificationProvider>
+				<ProfileProvider>
+					<Routes />
+				</ProfileProvider>
+			</NotificationProvider>
+		</ThemeProvider>
 	);
 }
